extraido el fetch y el renderizado de la data al componente especifico ListadoDePersonajes, y agregado un suspense sobre dicho componente para mostrar un skeleton mientras se completa el fetch. Asi ahora el componente principal puede renderizar sin esperar a que se complete el fetch
